requirements.txt
prompt_toolkit
python-dotenv==1.0.0
pandas
psycopg2-binary
